Initial commit to display student/instructor account type.
[FRONTEND] Modified ../topic/post.tpl to display user data field
storing the account type of either "student" or "instructor".

[BACKEND] Modified src/posts/user.js and src/user/create.js to dynamically
store accounttype as a user data field for accessing in the frontend
template.

PERSISTING BUG: It seems that the option to be student or instructor is not actually
working when a new user registers an account. So it's always defaulted
to "student" type, even if user clicks "instructor".