Fixed persisting bug with program not recognizing metadata storing account type.
Earlier bug was not storing the account type selected by the user
when registering with a new account. Fix: data was not being accessed
correctly. From src/controllers/authentication.js, determined that
data object stored 'account-type' as an element, not as an attribute.
Changed access pattern in src/user/create.js.